% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_mixing_ratio_data.R
\name{get_all_mixing_ratio_data}
\alias{get_all_mixing_ratio_data}
\title{get raw mixing ratio data}
\usage{
get_all_mixing_ratio_data(Run.Label)
}
\arguments{
\item{Run.Label}{Run from which to get the data (Indepedent or Dependent)}
}
\value{
Data frame containing raw budget data for further analysis
}
\description{
Get the raw data of the mixing ratios of all from the Mechanisms from a Run, that was processed at a Date
}

