% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytical_model_data.R
\name{get_analytical_model_data}
\alias{get_analytical_model_data}
\title{get raw data from analytical model variables}
\usage{
get_analytical_model_data(Mechanism, Run.Label, Date)
}
\arguments{
\item{Mechanism}{The mechanism to get the data frame}

\item{Run.Label}{Run from which to get the data (TD/TI)}

\item{Date}{The data processing date (ddmmyyyy)}

\item{Species}{The species whose budgets are being analysed}
}
\value{
Data frame containing raw data for further analysis
}
\description{
Get the raw data of the analytical model of Pusede:2014 for Mechanisms from a Run, that was processed at a Date
}

