% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_budget_data.R
\name{get_budget_data}
\alias{get_budget_data}
\title{get raw budget data}
\usage{
get_budget_data(Mechanism, Species, Run.Label, Date)
}
\arguments{
\item{Mechanism}{The mechanism to get the data frame}

\item{Species}{The species whose budgets are being analysed}

\item{Run.Label}{Run from which to get the data (TD/, I)}

\item{Date}{The data processing date (ddmmyyyy)}
}
\value{
Data frame containing raw budget data for further analysis
}
\description{
Get the raw data of the budgets of a Species in a Mechanisms from a Run, that was processed at a Date
}

