% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contour_axis_label.R
\name{get_contour_axis_label}
\alias{get_contour_axis_label}
\title{Axis Labels for Contour Plots}
\usage{
get_contour_axis_label(break.points, orig.data, digits)
}
\arguments{
\item{break.points}{Sequence at which the axis will be labelled}

\item{orig.data}{Original data frame containing the contour data for plotting}

\item{digits}{Number of digits to round the label to.}
}
\value{
Sequence of labels for axis labelling
}
\description{
Axis labels from original data to be used for labelling plot axes
}

