% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contour_data.R
\name{get_contour_data}
\alias{get_contour_data}
\title{Interpolate data of a species over each Temperature and NOx level for the species}
\usage{
get_contour_data(mechanism, species, dataframe)
}
\arguments{
\item{mechanism}{Chemical mechanism for which the interpolated data will be calculated}

\item{species}{The species for which contours will be analysed}

\item{dataframe}{The data frame included the raw data}
}
\value{
The data frame of mixing ratio data interpolated at each Temperature and NOx level
}
\description{
Interpolates the data frame of mixing ratio of input species over the range of NOx and Temperatures for a mechanism.
}

