% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net_budget_data.R
\name{get_net_budget_data}
\alias{get_net_budget_data}
\title{get data for net budget plot}
\usage{
get_net_budget_data(dataframe)
}
\arguments{
\item{dataframe}{The data frame containing full data}
}
\value{
The data frame with the data necessary for net budget plot
}
\description{
Select data for Net Budget plot, determine NOx-Condition and output mean Net Budget for each NOx-condition, Mechanism and Temperature
}

