% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOx_condition.R
\name{get_NOx_condition}
\alias{get_NOx_condition}
\title{Get NOx-Condition based on H2O2/HNO3 ratio Sillman:1995}
\usage{
get_NOx_condition(x)
}
\arguments{
\item{x}{Ratio of H2O2 to HNO3 mixing ratios}
}
\value{
The NOx-Condition associated with ratio
}
\description{
NOx-Condition is determined from H2O2/HNO3 ratio based on values from Sillman:1995.
}

