% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratio_data.R
\name{get_ratio_data}
\alias{get_ratio_data}
\title{get data of ratios of mixing ratios of 2 species over each NOx-Condition, every Mechanism and Run}
\usage{
get_ratio_data(dataframe, Numerator, Denominator)
}
\arguments{
\item{dataframe}{The data frame with the raw data of all mixing ratios}

\item{Numerator}{The species whose mixing ratios are the numerator or ratio}

\item{Denominator}{The species whose mixing ratios are the denominator or ratio}
}
\value{
Data frame containing ratios of mixing ratio data over each NOx-Conditions
}
\description{
Get the raw data of the budgets of a Species in a Mechanisms from a Run, that was processed at a Date
}

