% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_budget.R
\name{get_species_budget}
\alias{get_species_budget}
\title{Get budget allocated to reactions of species from raw data}
\usage{
get_species_budget(dataframe, Reactants, Absolute)
}
\arguments{
\item{dataframe}{The data frame with raw budget data}

\item{Reactants}{FALSE => full reaction returned, TRUE => Reactants are returned}

\item{Absolute}{TRUE => Absolute data returned, FALSE => Fractional contributions to total are returned}
}
\value{
Data frame with species budget data
}
\description{
Input of raw data frame is teased to output the reactions and their rates that contribute to the species production and consumption data. Reaction switch to output full reaction label or just reactants. Absolute or Fraction of total also calculated based on switch.
}

