% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalising_rates.R
\name{normalising_rates}
\alias{normalising_rates}
\title{Normalise production rates by total loss rate of emitted VOC}
\usage{
normalising_rates(value, Normalising.df, mechanism, temperature, run,
  NOx.condition)
}
\arguments{
\item{Normalising.df}{The df containing the values of the total loss rates of emitted VOC, grouped by Mechanism, Temperature, Run, NO.Condition}

\item{mechanism}{The mechanism to get the data frame}

\item{temperature}{The temperatures of the study}

\item{run}{The Run type (TD or TI)}

\item{NOx.condition}{The NOx.Condition (Low-NOx, Maximal-O3 or High-NOx)}

\item{Rate}{The reaction rate that needs to be normalised}
}
\value{
normalise The normalised rate
}
\description{
Divide the rate of a reaction ny total loss rate of all emitted VOC
}

