% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dO3_dT_comparison_ERA.R
\name{plot_dO3_dT_comparison_ERA}
\alias{plot_dO3_dT_comparison_ERA}
\title{Plot rate of change of O3 with temperature over the different NOx Conditions, overlaid with ERA summer values for particular region
Plots of the mean O3 mixing ratio in each Mechanism, facetted by NOx.Condition and Run}
\usage{
plot_dO3_dT_comparison_ERA(df, area)
}
\arguments{
\item{df}{Data frame with simulated data}

\item{area}{Subset of ERA grid for comparison}
}
\value{
The plot
}
\description{
Plot rate of change of O3 with temperature over the different NOx Conditions, overlaid with ERA summer values for particular region
Plots of the mean O3 mixing ratio in each Mechanism, facetted by NOx.Condition and Run
}

