% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net_budgets.R
\name{plot_net_budgets}
\alias{plot_net_budgets}
\title{Plot net Budgets 
Net Budget Plots of each Mechanism, facetted by NOx.Condition, linetypes are different runs}
\usage{
plot_net_budgets(df)
}
\arguments{
\item{df}{Data frame with plotting data}
}
\value{
The plot
}
\description{
Plot net Budgets 
Net Budget Plots of each Mechanism, facetted by NOx.Condition, linetypes are different runs
}

