% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratios.R
\name{plot_ratios}
\alias{plot_ratios}
\title{Plot ratios of mixing ratios 2 species
Plots of the ratios of the mixing ratios of 2 species in each Mechanism, facetted by NOx.Condition, linetypes are different runs}
\usage{
plot_ratios(df)
}
\arguments{
\item{df}{Data frame with plotting data}
}
\value{
The plot
}
\description{
Plot ratios of mixing ratios 2 species
Plots of the ratios of the mixing ratios of 2 species in each Mechanism, facetted by NOx.Condition, linetypes are different runs
}

