% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_budgets.R
\name{plot_species_budgets}
\alias{plot_species_budgets}
\title{Plots of Budgets of Species}
\usage{
plot_species_budgets(df, Absolute, Stacked)
}
\arguments{
\item{df}{The data frame with plotting data}

\item{Absolute}{TRUE => Plot absolute Rates, FALSE => Plot Fractional contributions}

\item{Stacked}{TRUE => Lines are stacked, FALSE => Lines are not-stacked}
}
\value{
Final plot
}
\description{
Plotting of budgets of the species. Switches for Absolute or Fractional data and Stacked or non-Stacked lines
}

