% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_budgets_stacked_area.R
\name{plot_species_budgets_stacked_area}
\alias{plot_species_budgets_stacked_area}
\title{Plots of Budgets of Species, stacked area}
\usage{
plot_species_budgets_stacked_area(df, Absolute)
}
\arguments{
\item{df}{The data frame with plotting data}

\item{Absolute}{TRUE => Plot absolute Rates, FALSE => Plot Fractional contributions}
}
\value{
Final plot
}
\description{
Plotting of budgets of the species. Switches for Absolute or Fractional data
}

