% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mixing_ratio_data.R
\name{read_mixing_ratio_data}
\alias{read_mixing_ratio_data}
\title{Read in mixing ratio data of a run type, output interpolated data for contour plot}
\usage{
read_mixing_ratio_data(run.label, spc, mechanisms)
}
\arguments{
\item{run.label}{The Run form which to extract data}

\item{spc}{The species for which to get contour data}

\item{mechanisms}{The mechanisms for which to get the data}
}
\value{
data frame of interpolated mixing ratio data at each Temperature and NOx level
}
\description{
Reads in csv data from NOx and T variations from a Run and adds the Run column with information about the Run label. Uses get_contour_data to interpolate the data for a specific species.
}

